+++ pkg/apis/kubernetes/v1/common.go
@@ -1,0 +1,63 @@
+package v1
+
+import (
+	"context"
+	"net/url"
+
+	"go.anx.io/go-anxcloud/pkg/api"
+	"go.anx.io/go-anxcloud/pkg/api/types"
+	"go.anx.io/go-anxcloud/pkg/utils/object/filter"
+)
+
+func endpointURL(ctx context.Context, o types.Object, apiPath string) (*url.URL, error) {
+	op, err := types.OperationFromContext(ctx)
+	if err != nil {
+		return nil, err
+	}
+
+	if op == types.OperationUpdate {
+		return nil, api.ErrOperationNotSupported
+	}
+
+	// we can ignore the error since the URL is hard-coded known as valid
+	u, _ := url.Parse(apiPath)
+
+	if op == types.OperationList {
+		helper, err := filter.NewHelper(o)
+		if err != nil {
+			return nil, err
+		}
+
+		filters := helper.BuildQuery().Encode()
+
+		if filters != "" {
+			query := u.Query()
+			query.Set("filters", filters)
+			u.RawQuery = query.Encode()
+		}
+	}
+
+	return u, nil
+}
+
+// commonRequestBody is embedded in the request body types of kubernetes resources
+type commonRequestBody struct {
+	// this allows us to provide the state as string
+	State string `json:"state,omitempty"`
+}
+
+// requestBody removes the request body for read operations on kubernetes resources
+func requestBody(ctx context.Context, br func() interface{}) (interface{}, error) {
+	op, err := types.OperationFromContext(ctx)
+	if err != nil {
+		return nil, err
+	}
+
+	if op == types.OperationCreate || op == types.OperationUpdate {
+		response := br()
+
+		return response, nil
+	}
+
+	return nil, nil
+}