+++ pkg/apis/kubernetes/v1/integration_test.go
@@ -1,0 +1,6 @@
+//go:build integration
+// +build integration
+
+package v1
+
+const isIntegrationTest = false // TODO: need a way to await nodepool deletion before deleting the cluster